fix sample exporting

+++ OTRExporter/AudioExporter.h
@@ -20,14 +20,14 @@ private:
     void WriteSequenceXML(ZAudio* audio);
     void WriteSampleBinary(ZAudio* audio);
     void WriteSampleXML(ZAudio* audio);
-    std::string GetSampleEntryReference(ZAudio* audio, SampleEntry* entry, std::map<uint32_t, SampleEntry*> samples);
-    std::string GetSampleEntryStr(ZAudio* audio, const std::pair<const uint32_t, SampleEntry*>& pair);
-    std::string GetSampleDataStr(ZAudio* audio, const std::pair<const uint32_t, SampleEntry*>& pair);
+    std::string GetSampleEntryReference(ZAudio* audio, SampleEntry* entry, std::map<uint64_t, SampleEntry*> samples);
+    std::string GetSampleEntryStr(ZAudio* audio, const std::pair<const uint64_t, SampleEntry*>& pair);
+    std::string GetSampleDataStr(ZAudio* audio, const std::pair<const uint64_t, SampleEntry*>& pair);
     void WriteEnvData(std::vector<AdsrEnvelope*> envelopes, BinaryWriter* writer);
     void WriteEnvData(std::vector<AdsrEnvelope*> envelopes, tinyxml2::XMLElement* xmlDoc);
-    void WriteSoundFontEntry(ZAudio* audio, SoundFontEntry* entry, std::map<uint32_t, SampleEntry*> samples,
+    void WriteSoundFontEntry(ZAudio* audio, SoundFontEntry* entry, std::map<uint64_t, SampleEntry*> samples,
                              BinaryWriter* writer);
-    void WriteSoundFontEntry(ZAudio* audio, SoundFontEntry* entry, std::map<uint32_t, SampleEntry*> samples,
+    void WriteSoundFontEntry(ZAudio* audio, SoundFontEntry* entry, std::map<uint64_t, SampleEntry*> samples,
                              tinyxml2::XMLElement* xmlDoc, const char* name);
     const char* GetMediumStr(uint8_t medium);
     const char* GetCachePolicyStr(uint8_t policy);
